requirements/base.txt
analytics-python==1.1.0
celery==3.1.18
Django==1.8.7
django-appconf==0.6
django-compressor==1.5
django_extensions==1.5.5
django-filter-0.11.0
django-libsass==0.5
django-oscar==1.1.1
django_simple_history==1.6.3
django-waffle==0.10.1
djangorestframework==3.2.3
djangorestframework-jwt==1.7.2
drf-extensions==0.2.8
edx-auth-backends==0.1.3
edx-ecommerce-worker==0.2.0
edx-rest-api-client==1.2.1
jsonfield==1.0.3
libsass==0.9.2
paypalrestsdk==1.11.4
premailer==2.9.2
pycountry==1.18
python-dateutil==2.4.2
pytz==2015.7
requests==2.8.1
suds==0.4
